requirements.txt
PyQt6>=6.4.0
PyMuPDF>=1.23.0
pypdf>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0
docx2pdf>=0.1.8
pdf2docx>=0.5.6
